#15: move v1 implementation to a dedicated file, change default port to 15151
