+++ app.py
@@ -4,7 +4,7 @@ from datetime import datetime, timedelta
 from flask import Flask, jsonify
 from flask.helpers import make_response
 from waitress import serve
-from vvspy.trip import get_trips
+import service.v1 as v1
 
 
 from_id = getenv("VVS_FROM", "5006333")
@@ -24,49 +24,14 @@ def check_time():
     return datetime.now() + timedelta(minutes=check_time_offset)
 
 
-def iso_ts(date):
-    return date.strftime('%Y-%m-%dT%H:%M:%SZ')
-
-
-def extract_data(connection):
-    departure_planed = connection.origin.departure_time_planned
-    departure_estimated = connection.origin.departure_time_estimated
-    departure_delta = int((departure_estimated - departure_planed).total_seconds() / 60)
-    arrival_planed = connection.destination.arrival_time_planned
-    arrival_estimated = connection.destination.arrival_time_estimated
-    arrival_delta = int((arrival_estimated - arrival_planed).total_seconds() / 60)
-
-    travel_time = int((arrival_estimated - departure_planed).total_seconds() / 60)
-    return {
-        "from_id": from_id,
-        "to_id": to_id,
-        "from": connection.origin.name,
-        "to": connection.destination.name,
-        "departure_planned": iso_ts(departure_planed),
-        "departure_estimated": iso_ts(departure_estimated),
-        "departure_delay": departure_delta,
-        "arrival_planed": iso_ts(arrival_planed),
-        "arrival_estimated": iso_ts(arrival_estimated),
-        "arrival_delay": arrival_delta,
-        "travel_time": travel_time,
-        "number": connection.transportation.number
-    }
-
-
-def find_trips():
-    trips = get_trips(from_id, to_id, limit=limit, check_time=check_time())
-    trip_data_list = [extract_data(trip.connections[0])
-                      for trip in trips if len(trip.connections) == 1]
-    return trip_data_list
-
-
 app = Flask(__name__)
 
 
+@app.route('/api/v1/', methods=['GET'])
 @app.route('/', methods=['GET'])
-def get_tasks():
+def get_trips_v1():
     try:
-        return jsonify({'trips': find_trips(), 'status': 200})
+        return jsonify({'trips': v1.find_trips(from_id, to_id, limit, check_time=check_time()), 'status': 200})
     except Exception as e:
         return jsonify({'message': str(e), 'status': 500}), 500
 
@@ -79,4 +44,4 @@ def after_request_func(data):
 
 
 if __name__ == "__main__":
-    serve(app, host="0.0.0.0", port=5000)
+    serve(app, host="0.0.0.0", port=15151)
